change condition to account for multivariate
